Modify popup and ingredients popup

+++ ZetaClient/ZetaClient/Entities/ModelIngredient.cs
@@ -9,7 +9,7 @@ namespace ZetaClient.Entities
 {
     public class ModelIngredient: AbstractEntity
     {
-        public float Grammage { get; set; }
+        public double Grammage { get; set; }
         public FrisbeeModel FrisbeeModel { get; set; }
         public Ingredient Ingredient { get; set; }
     }
